Move upload completion processor to its own package
